% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f1_championships}
\alias{f1_championships}
\title{F1 Championship Data}
\format{
\describe{
\item{number}{Rank of driver in the championship}
\item{driver}{first and last name of the driver}
\item{constructor}{name of the constructor (team) that designed the F1 car}
\item{points}{Number of points the driver earned that season}
\item{year}{Racing season}
}
}
\source{
\url{https://www.racing-statistics.com/}
}
\usage{
f1_championships
}
\description{
Dataset containing 1950-2022 F1 Championship results. Data is by racing
driver.
}
\keyword{datasets}
