% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f1_events_data}
\alias{f1_events_data}
\title{F1 Championship Data - Events}
\format{
\describe{
\item{number}{Event number of the season}
\item{date}{date of race}
\item{event}{name of the event}
\item{circuit}{name of the racing circuit}
\item{winning_driver}{Name of the driver that won the race}
\item{winning_constructor}{Name of the constructor that won the race}
\item{laps}{number of laps}
\item{time}{Duration of race}
\item{year}{Racing season}
}
}
\source{
\url{https://www.racing-statistics.com/}
}
\usage{
f1_events_data
}
\description{
Dataset containing 1950-2023 F1 Championship Event results.
}
\examples{

library(tidyverse)
data(f1_events_data)

f1_events_data \%>\%
  group_by(year) \%>\%
  summarize(n_events = n())


}
\keyword{datasets}
